pyproject.toml
[tool.poetry]
name = "light-control"
version = "0.1.0"
description = "control some lights!"
authors = ["k.hitchcock"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tapo = "^0.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
